+++ src/visual_editor/components/AttributeEdit.tsx
@@ -3,7 +3,10 @@ import { validAttributeName } from "dom-mutator";
 import React, { FC, useCallback, useState } from "react";
 import TextareaAutosize from "react-textarea-autosize";
 import { RxCross2, RxPlus, RxCheck } from "react-icons/rx";
-import { hoverAttributeName, selectedAttributeName } from "./lib/selectionMode";
+import {
+  hoverAttributeName,
+  selectedAttributeName,
+} from "../lib/selectionMode";
 
 export const IGNORED_ATTRS = [
   "class",
