+++ src/visual_editor/components/ElementDetails/DetailsRow.tsx
