+++ src/visual_editor/components/VisualEditorSection.tsx
